refactor(lsp): extract includes / commands / extension-boot from server.rs
#775 got server.rs to 1399 by moving tests + pure helpers into a `server/`
submodule. This continues that decomposition for the three concern groups
that were still inline — the "logic that crept in" the router doesn't need
to hold:

- server/extension_boot.rs  lazy extension_state() boot + invalidate
- server/includes.rs        resolve_and_upsert / goto_for_include /
                            hover_for_include
- server/command_dispatch.rs  executeCommand arg-parsing + extractToInclude
                            + preparePaste; the trait method is now a
                            one-line delegate to execute_command_dispatch()

The struct fields + document()/document_entry() are now pub(crate) so the
impl blocks can live in sibling modules (DocumentStore/DocumentEntry were
already pub(crate) from #775). server.rs is left as the LspClient trait, the
struct, and the LanguageServer router + lifecycle.

server.rs: 1399 → 660 lines (3644 before #775). 85 tests pass, clippy clean.
